Added profile project builder.
The builder is responsible for loading all profile application files contained in a project automatically.
So the user doesn't have to load such files manually via the UI.
A project's resource changes are observed. In case that a new profile application file was added, the builder is also responsible for loading that file.