Added initial implementation of interface EStereotypedObject.
This implementation has not been tested yet.

The goal is to get rid of the constructor's dependency to ProfileApplicationDecorator in order to avoid problems when using this implementation in the genmodel's code generation process.

The method "applyStereotype" still contains an approach (given in pseudo code) of how to be independent from editor part and thus independent from modelId and get rid of the constructor's dependency to ProfileApplicationDecorator.