Added workaround to run unit tests successfully again.
The problem was introduced with current implementation of EStereotypableObject.getApplicableStereotypes().
Stereotypes from different resource sets are retrieved both from profile registry and from the resource set of the model.
That's why the assertSame statements of stereotype comparison fail.

The first stereotype is from the profile registry resource set. This stereotype is applied to an EStereotypeableObject.
But the returned StereotypeApplication references a stereotype from the resource set of the model to which it is applied.

Workaround:
Get the stereotype again for comparison.
At the second request a decorator exists and the resource set of the model is used by the decorator.