Display signature name in KAMP
The (entity) name of signatures is displayed in the *.modificationsmarks file instead of the id (which has been belonging to the node in the tree, not to the modified signature!). This affects the labeling of manually added and computed signature nodes in the tree as well as the selection list for "Modify signature" nodes in the properties view.